oop/gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0   # JWT for CAMARA OIDC auth
python-multipart==0.0.9
openapi-spec-validator==0.7.1
